Create multi_Currency
Adds an array to the receipt where customers can manually enter a notation and exchange rate.
Fill in a notation, like USD or £. Fill in a rate manually, with a "." as decimal separator. Leave the array empty to not use the multi_currency function.
The second (or third... ) currency shows on every amount on the page. 
e.g.: €50 | 100 USD