Create better_Workorder
This work order template creates a workflow with the service module where the customer first receives a receipt as a proof of giving the item to the shop for repair. When the service is completed, the owner can choose to print the work order quote on A4 or to simply complete the sale.